admin abc

+++ src/vn/edu/nlu/fit/admin/servlet/RemoveCategoryServlet.java
@@ -8,6 +8,7 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 import java.io.IOException;
 import java.sql.SQLException;
 
@@ -15,22 +16,28 @@ import java.sql.SQLException;
 public class RemoveCategoryServlet extends HttpServlet {
 
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        String categoryIDString = request.getParameter("categoryID");
-        int categoryID;
-        try {
-            categoryID = Integer.parseInt(categoryIDString);
-        } catch (Exception e){
-            categoryID = 0;
-        }
-        CategoryService categoryService = new CategoryServiceImpl();
-        try {
-            if (categoryService.removeCategory(categoryID)>0){
-                response.setStatus(200);
-            } else {
-                response.setStatus(500);
-            }
-        } catch (SQLException e){
+        HttpSession session = request.getSession();
+        if (session.getAttribute("adminUser") == null) {
+            response.sendRedirect("/admin/login");
+        } else {
+            String categoryIDString = request.getParameter("categoryID");
 
+            int categoryID;
+            try {
+                categoryID = Integer.parseInt(categoryIDString);
+            } catch (Exception e) {
+                categoryID = 0;
+            }
+            CategoryService categoryService = new CategoryServiceImpl();
+            try {
+                if (categoryService.removeCategory(categoryID) > 0) {
+                    response.setStatus(200);
+                } else {
+                    response.setStatus(500);
+                }
+            } catch (SQLException e) {
+
+            }
         }
     }
 }
